+++ pdfimpose/__init__.py
@@ -17,7 +17,7 @@
 
 "Perform an imposition on the PDF file given in argument."
 
-VERSION = "2.1.1"
+VERSION = "2.2.0"
 
 
 class UserError(BaseException):